Revert "Moved dialogs classes into separate modules and some clean"
This reverts commit 903e8124d4e98f3811aed139c56fb8df5aabaa60.
